:pushpin: refactor components, add export functions in utilities folder
