+++ src/Utilities/index.js
@@ -1,0 +1,12 @@
+export * from './countDuplicates';
+export * from './dataForChartAccordingToYear';
+export * from './excelDateToJSDate';
+export * from './formatColumnsToTable';
+export * from './formatDataToCartesianChart';
+export * from './formatDataToTable';
+export * from './getItemsArray';
+export * from './getMonth';
+export * from './getYear';
+export * from './numberAmountToPesoFormat';
+export * from './selectLabelOption';
+export * from './totalSellsAccordingToMonth';
