+++ src/components/AreaChart.jsx
@@ -13,7 +13,7 @@ import {
 import { Line } from 'react-chartjs-2';
 import { Flex, Select, Stack, VStack } from '@chakra-ui/react';
 import { useDataContext } from '../context/DataContext';
-import { countDuplicates } from './ProcessDataToCharts';
+import { countDuplicates, dataForChartAccordingToYear, totalSellsAccordingToMonth } from '../Utilities';
 
 ChartJS.register(
   CategoryScale,
@@ -67,11 +67,13 @@ const labels = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 
 
 
 function AreaChart() {
-  const {chartAreaData, chartData} = useDataContext()
+  const {cartesianChartData, chartData} = useDataContext()
+  console.log('CARTESIAN DATA', cartesianChartData);
   const selectYear = chartData ? Object.keys(countDuplicates(chartData.año)) : ['Select year']
   const [select1, setSelect1] = useState('año')
   const [select2, setSelect2] = useState('año')
   const [yearDataSet1, setYearDataSet1] = useState([1, 2, 3])
+  console.log('YEAR DATASET', yearDataSet1);
   const [monthDataSet1, setMonthDataSet1] = useState([1, 2, 3])
   const [yearDataSet2, setYearDataSet2] = useState([1, 2, 3])
   const [monthDataSet2, setMonthDataSet2] = useState([1, 2, 3])
@@ -82,80 +84,6 @@ function AreaChart() {
   const handleChangeSelect2 = (event) => {
     setSelect2(event.target.value)
   }
-  const dataForChartAccordingToYear = (dataBase, yearSelected) => {
-    let data = dataBase.filter((item) => item["Año"] === yearSelected)
-    return data
-  }
-  const totalSellsAccordingToMonth = (array) => {
-    let january = 0
-    let february = 0
-    let march = 0
-    let april = 0
-    let may = 0
-    let june = 0
-    let july = 0
-    let august = 0
-    let september = 0
-    let october = 0
-    let november = 0
-    let december = 0
-
-    array.map((item) => {
-      if(item["Mes"] === "1"){
-        january += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "2"){
-        february += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "3"){
-        march += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "4"){
-        april += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "5"){
-        may += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "6"){
-        june += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "7"){
-        july += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "8"){
-        august += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "9"){
-        september += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "10"){
-        october += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "11"){
-        november += Math.floor(item[" Importe venta total "]) 
-      }
-      if(item["Mes"] === "12"){
-        december += Math.floor(item[" Importe venta total "]) 
-      }
-    })
-    let dataMonth = 
-    {
-      'January' : january,
-      'February' : february,
-      'March' : march,
-      'April' : april,
-      'May' : may,
-      'June' : june,
-      'July' : july,
-      'August' : august,
-      'September' : september,
-      'October' : october,
-      'November' : november,
-      'December' : december,
-    }
-    return dataMonth
-  }
-
   const data = {
     labels,
     datasets: [
@@ -178,7 +106,7 @@ function AreaChart() {
 
   useEffect(() => {
     const setYearData = async () => {
-      const dataArray = await dataForChartAccordingToYear(chartAreaData, select1)
+      const dataArray = await dataForChartAccordingToYear(cartesianChartData, select1)
       setYearDataSet1(dataArray)    
     }
     setYearData()
@@ -186,7 +114,7 @@ function AreaChart() {
 
   useEffect(() => {
     const setYearData = async () => {
-      const dataArray = await dataForChartAccordingToYear(chartAreaData, select2)
+      const dataArray = await dataForChartAccordingToYear(cartesianChartData, select2)
       setYearDataSet2(dataArray)    
     }
     setYearData()
