+++ src/components/BarChart.jsx
@@ -10,9 +10,8 @@ import {
     Legend,
   } from 'chart.js';
 import { Bar } from 'react-chartjs-2';
-
 import { useDataContext } from '../context/DataContext';
-import { countDuplicates } from './ProcessDataToCharts';
+import { countDuplicates, dataForChartAccordingToYear, totalSellsAccordingToMonth } from '../Utilities';
 
 ChartJS.register(
     CategoryScale,
@@ -62,7 +61,7 @@ ChartJS.register(
   const labels = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'];
 
 function BarChart() {
-    const {chartAreaData, chartData} = useDataContext()
+    const {cartesianChartData, chartData} = useDataContext()
     const selectYear = chartData ? Object.keys(countDuplicates(chartData.año)) : ['Select year']
     const [select1, setSelect1] = useState('año')
     const [select2, setSelect2] = useState('año')
@@ -77,80 +76,6 @@ function BarChart() {
     const handleChangeSelect2 = (event) => {
       setSelect2(event.target.value)
     }
-    const dataForChartAccordingToYear = (dataBase, yearSelected) => {
-      let data = dataBase.filter((item) => item["Año"] === yearSelected)
-      return data
-    }
-    const totalSellsAccordingToMonth = (array) => {
-      let january = 0
-      let february = 0
-      let march = 0
-      let april = 0
-      let may = 0
-      let june = 0
-      let july = 0
-      let august = 0
-      let september = 0
-      let october = 0
-      let november = 0
-      let december = 0
-  
-      array.map((item) => {
-        if(item["Mes"] === "1"){
-          january += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "2"){
-          february += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "3"){
-          march += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "4"){
-          april += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "5"){
-          may += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "6"){
-          june += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "7"){
-          july += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "8"){
-          august += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "9"){
-          september += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "10"){
-          october += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "11"){
-          november += Math.floor(item[" Importe Coste total "]) 
-        }
-        if(item["Mes"] === "12"){
-          december += Math.floor(item[" Importe Coste total "]) 
-        }
-      })
-      let dataMonth = 
-      {
-        'January' : january,
-        'February' : february,
-        'March' : march,
-        'April' : april,
-        'May' : may,
-        'June' : june,
-        'July' : july,
-        'August' : august,
-        'September' : september,
-        'October' : october,
-        'November' : november,
-        'December' : december,
-      }
-      return dataMonth
-    }
-  
     const data = {
       labels,
       datasets: [
@@ -173,7 +98,7 @@ function BarChart() {
   
     useEffect(() => {
       const setYearData = async () => {
-        const dataArray = await dataForChartAccordingToYear(chartAreaData, select1)
+        const dataArray = await dataForChartAccordingToYear(cartesianChartData, select1)
         setYearDataSet1(dataArray)    
       }
       setYearData()
@@ -181,7 +106,7 @@ function BarChart() {
   
     useEffect(() => {
       const setYearData = async () => {
-        const dataArray = await dataForChartAccordingToYear(chartAreaData, select2)
+        const dataArray = await dataForChartAccordingToYear(cartesianChartData, select2)
         setYearDataSet2(dataArray)    
       }
       setYearData()
