+++ src/components/DonutChart.jsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from 'react'
 import { Chart as ChartJS, ArcElement, Tooltip, Legend } from 'chart.js';
 import { Doughnut } from 'react-chartjs-2';
 import { useDataContext } from '../context/DataContext';
-import { selectLabel } from './ProcessDataToCharts';
+import { selectLabelOption } from '../Utilities';
 
 ChartJS.register(ArcElement, Tooltip, Legend);
 export const options = {
@@ -70,8 +70,8 @@ function DonutChart() {
   }
 
   useEffect(() => {
-    const labelArray = chartData && select ? Object.keys(selectLabel(chartData,select)) : ['option-1','option-2', 'option-3']
-    const dataArray = chartData && select ? Object.values(selectLabel(chartData,select)) : [1, 2, 3]
+    const labelArray = chartData && select ? Object.keys(selectLabelOption(chartData,select)) : ['option-1','option-2', 'option-3']
+    const dataArray = chartData && select ? Object.values(selectLabelOption(chartData,select)) : [1, 2, 3]
     setLabelsChart(labelArray)
     setDataChart(dataArray)    
   }, [select])
